+++ examples/data-sources/nxos_bridge_domain/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_bridge_domain" "example" {
+  fabric_encap = "vlan-10"
+}