+++ examples/data-sources/nxos_vrf/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_vrf" "example" {
+  name = "VRF1"
+}