+++ examples/data-sources/nxos_vrf_container/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_vrf_container" "example" {
+  name = "VRF1"
+}