+++ examples/resources/nxos_bridge_domain/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_bridge_domain.example "sys/bd/bd-[vlan-10]"