+++ examples/resources/nxos_bridge_domain/resource.tf
@@ -1,0 +1,5 @@
+resource "nxos_bridge_domain" "example" {
+  fabric_encap = "vlan-10"
+  access_encap = "unknown"
+  name         = "VLAN10"
+}