+++ examples/resources/nxos_vrf/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_vrf.example "sys/inst-[VRF1]"