+++ examples/resources/nxos_vrf/resource.tf
@@ -1,0 +1,4 @@
+resource "nxos_vrf" "example" {
+  name        = "VRF1"
+  description = "My VRF1 Description"
+}