+++ examples/resources/nxos_vrf_container/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_vrf_container.example "sys/ipv4/inst/dom-[VRF1]"