+++ examples/resources/nxos_vrf_container/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_vrf_container" "example" {
+  name = "VRF1"
+}