+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-## 0.2.2 (unreleased)
+## 0.2.2
 
 - Add nxos_ipv4_interface resource and data source
 - Add nxos_ipv4_interface_address resource and data source
+- Add nxos_vrf resource and data source
+- Add nxos_vrf_container resource and data source
+- Add nxos_bridge_domain resource and data source
 
 ## 0.2.1